Refactor yield curve construction and handling
This commit streamlines yield curve handling by using the qlab-math crate to perform interpolation instead of the previously implemented linear interpolation contained within the qlab-termstructure crate. Consequently, this led to the deletion of `grid_point.rs` and `linear_interpolation.rs`. This refactoring simplifies the termstructure codebase, making it more reusable, and in aligning with best practices, it improves code modularity and maintainability.