Refactor interpolation interface and use in qlab-termstructure
This commit overhauls the interpolation interface in qlab-math and updates the yield curve model in qlab-termstructure to utilize the new interface. The private 'InterpolatorInner' trait has been removed and an abstract 'Method' trait has been introduced, changing the way 'Interpolator' objects fit the structured data. Additionally, relevant testing so as to ensure the implemented changes function as expected.