Add public holidays calendars for Ukraine, UK, and US
Added implementation of public holidays for Ukraine, the UK, and the US in the third-party calendar module. Including various markets and special closures days. Calendar functionality was expanded to indicate business days specific to these countries.  This will aid in accurate business day calculations for these markets. Test cases were also added to validate the accuracy of the holidays and business days implemented.